SmoothCoreExecutorServiceBuilder use short

+++ misty-smooth-core/src/main/java/org/misty/smooth/core/constant/ThreadPoolArgument.java
@@ -7,31 +7,31 @@ public class ThreadPoolArgument {
     public static class CoreSize {
         public static final String KEY = PREFIX + ".core-size";
 
-        public static final int PRESET = 2;
+        public static final short PRESET = 2;
 
-        public static final int MIX = PRESET;
+        public static final short MIX = PRESET;
 
-        public static final int MAX = 256;
+        public static final short MAX = 256;
     }
 
     public static class MaxSize {
         public static final String KEY = PREFIX + ".max-size";
 
-        public static final int PRESET = 1024;
+        public static final short PRESET = 1024;
 
-        public static final int MIX = CoreSize.MAX;
+        public static final short MIX = CoreSize.MAX;
 
-        public static final int MAX = 4096;
+        public static final short MAX = 4096;
     }
 
     public static class AliveSecond {
         public static final String KEY = PREFIX + ".alive-second";
 
-        public static final int PRESET = 60;
+        public static final short PRESET = 60;
 
-        public static final int MIX = PRESET;
+        public static final short MIX = PRESET;
 
-        public static final int MAX = 3600;
+        public static final short MAX = 3600;
     }
 
     public static class NamePrefix {
@@ -43,11 +43,11 @@ public class ThreadPoolArgument {
     public static class Rotation {
         public static final String KEY = PREFIX + ".rotation";
 
-        public static final int PRESET = 999;
+        public static final short PRESET = 999;
 
-        public static final int MIX = 1;
+        public static final short MIX = 1;
 
-        public static final int MAX = PRESET;
+        public static final short MAX = PRESET;
     }
 
 }
